adding cars app and starting to build infrastructure for BS5

+++ cars/signals.py
@@ -24,6 +24,6 @@ def auto_delete_thumbnail_file_on_change(sender, instance, **kwargs):
         return False
 
     new_file = instance.thumbnail
-    if not old_file == new_file:
+    if old_file.name and not old_file == new_file:
         if os.path.isfile(old_file.path):
             os.remove(old_file.path)
